# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST API for Ecomemerce app built with express.js and psql",
    "version": "0.1",
    "title": "Ecommerce REST API"
  },
  "paths": {
...
    "/cart/newcart": {
      "post": {
        "summary": "Creates new cart for authenticated user",
        "consumes": [
          "application/json"
        ],
        "parameters": [{
          "in": "body",
          "name": "body",
          "required": false,
          "schema": {
            "$ref": "#/definitions/Model0"
          },
          "x-examples": {
            "application/json": "{\n        \"product_name\": \"iphone\",\n        \"category\": \"technology\",\n        \"description\": \"iphone 12 pro max 256gb\",\n        \"price\": 800\n}"
          }
        }],
        "responses": {
          "201": {
            "description": "Returns newly created cart",
            "schema": {
              "type": "string"
            }
          }
        },
        "tags": [
          "cart"
        ]
      }
    },
    "/cart/items": {
      "post": {
        "summary": "Adds items to the users cart",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "body",
          "required": false,
          "schema": {
            "$ref": "#/definitions/Model9"
          },
          "x-examples": {
            "application/json": "[{\n    \"product_name\": \"Vase\",\n    \"quantity\": 3\n},\n{\n    \"product_name\": \"LG TV\",\n    \"quantity\": 1\n}]"
          }
        }],
        "responses": {
          "201": {
            "description": "Returns products added to the cart",
            "schema": {
              "$ref": "#/definitions/Model18_Array"
            }
          }
        },
        "tags": [
          "cart"
        ]
      },
      "delete": {
        "summary": "Deletes user cart based on authenticated user",
        "consumes": ["application/json"],
        "produces": ["text/html"],
        "parameters": [],
        "responses": {
          "204": {
            "description": "Confirms cart deletion",
            "schema": {
              "type": "string"
            }
          }
        },
        "tags": [
          "cart"
        ]
      }
    },
    "/cart/items/{productId}": {
      "put": {
        "summary": "Updates quantity of item in the user cart",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "body",
          "required": false,
          "schema": {
            "$ref": "#/definitions/Model3"
          },
          "x-examples": {
            "application/json": "{\n    \"quantity\": 4\n}"
          }
        }],
        "responses": {
          "200": {
            "description": "Confirms cart update",
            "schema": {
              "$ref": "#/definitions/Model12_Array"
            }
          }
        },
        "tags": [
          "cart"
        ]
      },
      "delete": {
        "summary": "Removes item from user cart",
        "consumes": ["application/json"],
        "parameters": [  {
            "name": "productId",
            "in": "path",
            "description": "Id of chosen product. Example values for path are: '3' and '1'",
            "required": true,
            "type": "integer"
          }],
        "responses": {
          "204": {
            "description": "Confirms item deleted from the user cart",
            "schema": {
              "type": "string"
            }
          }
        },
        "tags": [
          "cart"
        ]
      }
    },
...
    "/cart/checkout": {
      "post": {
        "summary": "Checks payment info for user and produces order based on user cart contents",
        "consumes": ["application/json"],
        "parameters": [{
          "in": "body",
          "name": "body",
          "required": false,
          "description": "Body requires card_number and expiry. Example: '4431562178013497' and '2024-11-14'",
          "schema": {
            "$ref": "#/definitions/Model24_Array"
          },
          "x-examples": {
            "application/json": "{\n    \"card_number\": 4431562178013497,\n    \"expiry\": \"2024-11-14\"\n}"
          }
        }],
        "responses": {
          "201": {
            "description": "Payment accepted, order generated and checkout confirmed",
            "schema": {
              "type": "string"
            }
          }
        },
        "tags": [
          "cart"
        ]
      }
    }
  },
  "definitions": {
    "Model0": {
      "properties": {
        "product_name": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Model3": {
      "properties": {
        "product_name": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Model9": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Array"
      }
    },
...
    "Model12_Array": {
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "product_name": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "type": "string"
        }
      }
    },
...
    "Array": {
      "properties": {
        "product_name": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Model18_Array": {
      "properties": {
        "cart_id": {
          "type": "integer",
          "format": "int32"
        },
        "product_id": {
          "type": "integer",
          "format": "int32"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Model24_Array": {
      "properties": {
        "card_number": {
          "type": "integer",
          "format": "varchar(20)"
        },
        "expiry": {
          "type": "string"
        }
      }
    },
...
  }
}